Sign out button, authReducer  bug, comment/labels for code
Created button which gives user possibility to sign out when  logged in.
Fixed bug in auth reducer that caused code set reverted auth state (false when logged in, and true when not logged in).
I know comments in code have to describe actions, not be title of parts of code but although I have added "comment-labels" to clearly divide parts of code for 1. Imports 2. Styled 3. React component 4. Redux maps/connect